api-service/app/usuario/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==1.4.46
pydantic==2.11.1
python-dotenv==0.21.1
databases==0.7.0
alembic==1.10.3
psycopg2-binary==2.9.6
pytest==7.2.2  
email-validator
passlib
pyjwt
